Message Queues, Named Pipes and Sending Signals
A controller device scenario where the device generates values (to mimic
data such as temperature) and sends it via the message queue to the
controller, if value generated is greater than user inputted  threshold
controller sends a signal to the cloud and the device they both print
out an error message. cloud communicates with controller via a named
pipe.